Fixing CI
