+++ web/app/types.ts
@@ -1,0 +1,11 @@
+export const ElementTypes = {
+  ELEMENT: 0,
+  COMPONENT: 1,
+  SLOT: 2,
+  TEMPLATE: 3,
+} as const;
+
+export const NodeTypes = {
+  ELEMENT: 1,
+  ATTRIBUTE: 6,
+} as const;
